fix: rename deploy related entities
Overview:
DEP-105
Throughout the code, we call the result of dynamo build with different terms:
service, components, graph, pipeline. Coalesce on one term: dynamograph_package.
Also, nim should be renamed to component.
