+++ deploy/sdk/src/dynamo/sdk/cli/cli.py
@@ -24,8 +24,8 @@ from rich.console import Console
 
 from dynamo.sdk.cli.deployment import app as deployment_app
 from dynamo.sdk.cli.deployment import deploy
+from dynamo.sdk.cli.dynamo_graph import build, get
 from dynamo.sdk.cli.env import env
-from dynamo.sdk.cli.pipeline import build, get
 from dynamo.sdk.cli.run import run
 from dynamo.sdk.cli.serve import serve
